refactor: split zipfile test util into its own file.
